put ActionController test compatibility into separate file
