+++ spec/spec_helper.rb
@@ -45,72 +45,6 @@ end
 # in ./support/ and its subdirectories.
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
-module ActionControllerTestCaseBehaviorCompatibility
-  def get(action, **args)
-    if Rails.version < '5.0.0'
-      parameters, session, flash = self.__strongbolt_params_compatibility(args)
-      super(action, parameters, session, flash)
-    else
-      super
-    end
-  end
-
-  def post(action, **args)
-    if Rails.version < '5.0.0'
-      parameters, session, flash = self.__strongbolt_params_compatibility(args)
-      super(action, parameters, session, flash)
-    else
-      super
-    end
-  end
-
-  def patch(action, **args)
-    if Rails.version < '5.0.0'
-      parameters, session, flash = self.__strongbolt_params_compatibility(args)
-      super(action, parameters, session, flash)
-    else
-      super
-    end
-  end
-
-  def put(action, **args)
-    if Rails.version <= '5.0.0'
-      parameters, session, flash = self.__strongbolt_params_compatibility(args)
-      super(action, parameters, session, flash)
-    else
-      super
-    end
-  end
-
-  def delete(action, **args)
-    if Rails.version < '5.0.0'
-      parameters, session, flash = self.__strongbolt_params_compatibility(args)
-      super(action, parameters, session, flash)
-    else
-      super
-    end
-  end
-
-  def head(action, **args)
-    if Rails.version < '5.0.0'
-      parameters, session, flash = self.__strongbolt_params_compatibility(args)
-      super(action, parameters, session, flash)
-    else
-      super
-    end
-  end
-
-  def __strongbolt_params_compatibility(**args)
-    parameters, session, flash = args
-    parameters ||= {}
-    return parameters.delete(:params), session, flash
-  end
-end
-
-module ActionController::TestCase::Behavior
-  prepend ActionControllerTestCaseBehaviorCompatibility
-end
-
 RSpec.configure do |config|
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   # config.fixture_path = "#{::Rails.root}/spec/fixtures"
